# backend/docs/swagger.yaml
definitions:
  models.ErrorMsg:
    properties:
      error:
        type: string
    type: object
...
  models.Grammar:
    properties:
      description:
        type: string
      example:
        items:
          $ref: '#/definitions/models.GrammarExample'
        type: array
      id:
        type: integer
    type: object
  models.GrammarExample:
    properties:
      chinese:
        type: string
      example:
        type: string
    type: object
...
  models.ReadingMaterial:
    properties:
      chinese:
        type: string
      content:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  models.SearchResult-models_Grammar:
    properties:
      count:
        type: integer
      page:
        type: integer
      pageSize:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.Grammar'
        type: array
    type: object
...
  models.SearchResult-models_ReadingMaterial:
    properties:
      count:
        type: integer
      page:
        type: integer
      pageSize:
        type: integer
      results:
        items:
          $ref: '#/definitions/models.ReadingMaterial'
        type: array
    type: object
...
  models.SuccessMsg:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: Easy Japanese API
  version: "0.1"
paths:
...
  /api/grammar/add:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Grammar'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Add grammar
      tags:
      - globalDictOp
  /api/grammar/delete:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Grammar deleted
          schema:
            $ref: '#/definitions/models.SuccessMsg'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Delete grammar
      tags:
      - globalDictOp
  /api/grammar/edit:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Grammar'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Edit grammar
      tags:
      - globalDictOp
  /api/grammar/get:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: RPP
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Grammar'
            type: array
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Browse all grammars
      tags:
      - globalDictOp
  /api/grammar/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: RPP
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchResult-models_Grammar'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Search among all grammars
      tags:
      - globalDictOp
...
  /api/reading-material/add:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ReadingMaterial'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Add reading material
      tags:
      - globalDictOp
  /api/reading-material/delete:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Delete reading material
      tags:
      - globalDictOp
  /api/reading-material/edit:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.ErrorMsg'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Edit reading material
      tags:
      - globalDictOp
  /api/reading-material/get:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: RPP
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReadingMaterial'
            type: array
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Browse all reading materials
      tags:
      - globalDictOp
  /api/reading-material/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Results per page
        in: query
        name: RPP
        type: integer
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SearchResult-models_ReadingMaterial'
        "400":
          description: Database error
          schema:
            $ref: '#/definitions/models.ErrorMsg'
      security:
      - JWTAuth: []
      summary: Fuzzy search in all reading materials
      tags:
      - globalDictOp
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"